TP-342: Refactor GoodsNomenclatureIndents for phantom headers.
In cases where there are phantom headers right below a chapter
(regardless of whether the good is directly within its lineage)
some goods are still given a "00" indent, despite being 3
levels into the hierarchy.

This commit checks for phantom headers and handles this edge
case in the importer.

With this indent is more and more abstracted away from
the depth attribute of a materialized path node. Therefore
it is less and less reasonable to infer the indent from
the depth without incurring large numbers of queries.

To handle this this commit also adds the indent column to
the GoodsNomenclatureIndent table so that the TARIC defined
indent (not the tree depth) can be stored here for easier
access.