Origin destination codes on measure type are not redundant
We previously thought that the origin destination
code on a measure type was always equal to the
trade movement code, but it turns out not to be
the case for national measures.

So we now store the origin destination code as an
explicit field, and when migrating the database
handle those measure types where it is different
explicitly.