Handle errors in legacy data in indent importer
We discovered that a number of commodity codes
don't obey the rule "my indent is +1 of my parent
indent". We think these are errors as there is a
small nubmer of them and no discernible pattern to
them, so we have udpated the indent importer to
detect their specific SIDs and correct the data.