Handle partially-loaded elements in XML parsing
The `etree` streaming API will sometimes emit start
events for elements without their children if
passed a IO object or the Pull mecahnism is used.
This is because the API will consume input in
blocks and return as many events as it can, rather
than waiting for elements to be complete.

So we are not able to check what elements an
element has when handling a start event. To get
around this, this commit removes the need for
looking at if an element is a leaf node and
instead introduces extra state fields to keep
track of if the parser is currently parsing an
element or not, which is sufficient for selecting
the right parser. (This should also remove the
limitation on only having a single level of
nesting.)