TP-340: Importers for origins and successors
These fight with the import system a little bit
because there's not enough information from the
XML alone to unambiguously look up the derived or
absorbed goods nomenclature. So instead we have to
use the resolved_links structure, which works fine
for as long as the links are resolved in order.