Add missing validations for geographical areas
There are a number of ways to break GA3 that we
haven't accounted for in our current business
rules (although it seems our downstream systems
are only able to catch some of these).

Most of this code is lifted directly from the
additional codes application... which makes me
wonder if it would be worth abstracting out a
common pattern here for description-period-based
business rules.