TP-398: Relax geo area validations to match seed file data
Two current validations cause the seed file to break for Geographical
Areas and its associated models.

The first is that a group member must also be a member of any groups
parents. This is assumed to be an accurate validation - however the
seed file does fail against it. With the data given by the seed
file it is likely this is fixed as the importer progresses through
the seed file. If it is not then all the data required to fix the
relationships is still provided. As a result this validation has
been disabled so that it can be re-enabled and checked against once
the seed file is imported.

The second validation is one to ensure validity ranges do not overlap
for GeographicalAreaDescriptions. As these descriptions only ever have
a start date it is impossible to not violate this check and so the
constraint has been removed.