Fix HMRC SDES notification payload JSON
The Upload model generates a JSON payload ready to be sent to the HMRC
SDES API, but had an error where it passed a UUID object instead of a
string.

This commit fixes that error and also sets a default value for the `now`
argument to `Upload.notify_hmrc()` for convenience.