Change the default column to be explicit rather than implict end date
By using the "real" end date we accidentally allow
ourselves to violate the constraint that if a
measure is end-dated then it must have a justification
regulation attached, which won't be the case if
the measure would expire natually.

So instead use the "explicit" end date as the
validity end date and include a separate column
for the "real" end date which we can use for
looking at whether we want to end date the measure
or not.