TP-371: Partition workbaskets into transactions.
Initially it was assumed that, as they are the visualised
unit of work for the user, WorkBaskets should be the unit of
work for the backend models. However it has become apparent
that this does not work with the what is required by HMRC.
Instead the Transaction is required to be the unit of work
within the backend. This is because even though the
WorkBasket is where all the work is made, it must be split
and transmitted as transactions to the clients. As a result
a WorkBasket must be split into several smaller units of
work - Transactions.

Given the centrality of WorkBaskets this is a rather fundamental
change. All TrackedModels now must refer to Transactions instead
of WorkBaskets and some of the underlying queries must adapt
accordingly. Furthermore during the process of this refactor it
was found that business rules were quite repetitive and would be
best served by being split into inheritable classes that were
solely run by the transaction and not by the models themselves.

Some tests, the BDD tests and one exporter test, have been
skipped as they are currently broken. But given current priorities
they will be fixed in a later commit.