TP-462: Add a distributed cache for the nursery and fix handlers
The TaMaTo app is most likely going to run on a distributed system
once deployed. As it stands the cache for both the django system
and importer nursery rely on localised stateful systems (e.g. a file
based or in memory cache).

This commit adds redis as the default cache for django, and adds
several cache "engines" for the nursery, including a redis engine
which can be used. The default is left as the in memory engine.

On top of this the importing system has been improved to better
handle foreign key references - especially those handlers which
rely on foreign keys for identifying fields. All new models classed
as "current" are also now cached  based on their identifying fields,
storing both the primary key and the class name in the cache. This
is to reduce the number of queries required when importing models
as the cache can be used instead of the database.

Measures have been updated to better reflect their effective end date
(this can now be generated within a queryset reducing query number again).

Database indexes have been added for many fields - focusing on fields
used in validator queries - this reduces query time significantly.