TP-462: Allow the importer to run in a distributed system.
To enable the importer to run in a distributed manner it needs
access to the chunks of XML to import through a shared system.
The tasks to run the import must also be shared in a similar way.

This commit adds importer database models which hold chunks of XML
that all instances of the system can access. These chunks are split
firstly on size (currently 50MB is the default). However if the file
being imported can be done so in an asynchronous manner it is then
also split on record code and (for measures and commodities) chapter
heading. These chunks can then be imported simultaneously
(asynchronous imports are almost solely intended for seed files).

The function for running an import has been converted into a Celery
task so this can also be handled in a distributed manner.

The importer caching system, for caching individual models, has been
moved to the nursery - so as to de-pollute the models system.
Some business rules and handlers have been further updated.