TP-451: Update business rules to be inline with TP-398
TP-371 included an overhaul of the business rules system to make it
more generic an inheritable. Whilst a sensible move, it also rendered
much of the work in TP-398 too difficult to merge. The end result was
a refactored and improved business rules system, with outdated business
rules.

This commit updates the business rules to be inline with TP-398 whilst
also conforming to the new business rules system. It also adds the
business rules to the relevant models so that they are used when
adding transactions.