TP-542: Update frontend views to be inline with current relations.
Historically descriptions didn't have a named `related_name`. This has
changed recently and so the frontend queries were prefetching the
wrong related name and erroring as a result.

This commit updates the queries to use the correct related_name for
descriptions - it also improves the caching setup for fetching the
latest descriptions. This may need to be made more generic and more
flexible in the near future.