Allow disabling SSO
When running locally, during development, it may be useful to disable
SSO and avoid requiring an internet connection.

This change adds an environment variable `SSO_ENABLED` which switches
on the SSO authentication flow if truthy (eg "on", or "true"), and
disables it if falsy - falling back to Django's auth ModelBackend, which
authenticates users in the database.