Handle successors deleted after their replacing code is unterminated
The EU TARIC we have recently processed contains a
number of instances of codes being end dated and
then subsequently updated to remove the end date.
In this case, the code is first updated and then
the successor removed. This breaks the import of
the successor because the importer expects an end
date to be on the replacing code in order to look
up the correct absorbing code, but it has just
been removed.

Instead, in this specific case we now look up the
absorbed link using the previous version of the
successor. As this is a deletion, the data should
in the XML should not have changed since the last
update so this is safe to do.