Uplift measure creation pattern for production use
This pattern was produced during the Transition
Period data migration as a simple way to create a
new measure along with all of the associated models
from a single function call.

This commit uplifts this to use the proper TaMaTo
database, save the models and remove some complex
logic around dates that is no longer needed.

This pattern is still useful for migrations and
may also be useful as a way of creating measures
from the UI.