TP-575: Add an Activity Stream endpoint for Data Workspace.
It is a requirement that the tariff data be open and as accessible as possible.
Part of this is to make the data available on Data Workspace, and the advised
route to do so is to expose an endpoint that provides all the current tariff
data in an ActivityStream format.

This commit adds an endpoint in the exporter package which formats all the
latest TrackedModels (paginated by 50) in the ActivityStream format ready
for consumption by DataWorkspace.