TP-609: Include objects up to a given transaction in .current
When checking actual current data `.current` works adequately.
However in many cases, especially the business rules this isn't
enough. In these cases the normal requirement is to check the
most current data for an object, or the data for the object in
the current transaction, or the data for the object in any
preceding transaction that is within the same workbasket as the
current transaction. Whichever has the greatest ID (so is most
recent) needs to be returned.

This commit adds a new QuerySet method called `current_as_of`
which takes a transaction and executes a query that satisfies
this requirement.