Patch 2 for showing only current members in geo members tab
The previous patch would find only the current members (via the
geographical area table), however as the query was proxied through
memberships (a separate table) it still resulted in duplicate
members being shown.

This commit introduces a new method which builds a list of members
based on the memberships being current - removing duplicates.