Add measure condition(s) string generation
When measures are listed in the UI, one of the column items is a list of conditions
associated with the measure.

This commit adds convenience methods and properties to the Measure model and associated
models to easily construct these lists.

This code was translated from the PHP prototype, which may not be complete or correct.