TP-643: Update current tolatest approved Tracked models.
The current terminology, especialy "current" is confusing.
What is current is dependent on the context of the
situation and therefore the query will change at times.
As a result `.current` is being renamed to `.latest_approved`
`.current_deleted` is being renamed to `.latest_deleted`.
`.current_as_of` is being renamed to `.approved_up_to_transaction`
and `.approved_or_in_transaction` has been deleted as it
is being superceded by `.approved_up_to_transaction`.

Furthermore, currently when CDS accepts data from TaMaTo the relevant
Tracked Models aren't flagged as the latest approved objects.
This only happens in the niche scenario when they are loaded
into the system pre-approved (which should only happens on
initial seed loads).

This commit adds an update to all the tracked models once
a workbasket is confirmed by CDS, flagging all the confirmed
models as the most latest approved versions.