TP-581: Linked model updates check business rules
Many business rules check that a condition holds only if the main model is modified, but
those rules could be violated if a linked model is updated. For example, ACN2 asserts:

> The referenced additional code type must exist and have as application code
> “non-Meursing” or “Export Refund for Processed Agricultural Goods".

But if the AdditionalCodeType is updated after the AdditionalCode is created and the
application code is changed, the rule is not checked again.

This commit adds a concept of "indirect business rules". A model may list business rules
that should be checked against any linked models that they apply to when it is updated.

The method of checking business rules is also modified to handle these indirect rule
triggers.

The business rule violation exception class has also been modified - each business rule
now automatically contains a specific violation subclass, eg ACN2.Violation. A violation
stores a reference to the violating model.