TP-640: Replace inappropriate parent nodes.
In some cases new Goods Nomenclature Indents are inserted in between
old indents and their children. When this happens it is important to
update the tree by migrating these children under the new indent.

This commit introduces an addition to the GoodsNomenclatureIndentHandler
post_save method, finding inappropriate parents of possible child nodes
and replacing them with the new indent.