TP-543: Update descriptions queries and fix business rules context.
Currently described models don't have an obvious single mechanism for
fetching related descriptions. `.descriptions` would be the logical
method, however this only provides descriptions directly attached to
the given version, which is innacurate.

This commit introduces a generic `.get_descriptions` method on the
`TrackedModel` class so all described `TrackedModels` can automatically
fetch all related descriptions. This method optionally takes a
`Transaction` which triggers the query to use `approved_up_to_transaction`
by default it will use `latest_approved`.

In this process it was discovered that business rules currently don't pay
attention to the current transaction - only the given models transaction.
As business rules now often cross multiple transactions, including
transactions which were made after the given model, it is important
that the correct transaction is provided to the business rules.

This commit therefore also introduces a transaction argument when
initialising a business rule so it can make the correct query.