Run business rules after every imported transaction
Previously the clean method on workbaskets was not
implemented, so there was no validation performed
on import.

But also, running against the entire workbasket is
not what we want – as per ADR 12, validations need
to be performed after every transaction.

This commit changes our import validation to run
business rules after every transaction, but also
implements a clean method for a workbasket as well
(it will be useful later).

Notably we also no longer halt an import run when
an error is found and instead just print an error.
The reason for this is that when importing we are
normally importing well-formed XML and business
rule violations are (at the moment) normally an
indication that our implementations are wrong.
Plus, it's much easier to debug data when it is in
the database rather than in the XML files. So this
gives us the option to do that more easily.

When we come around to implementing receipt
parsing and marking individual transactions as
errored in TP-452, we can change this behaviour
and as well as logging an error we can mark the
individual transaction as being in an error state.