TP-669: Show workbasket changes on index and add submit button.
Instead of showing a list of workbaskets on the index page it has been
decided to show the single active workbasket and its list of changes instead.
This is coupled with a button to submit the workbasket.

This commit updates the index view to provide a single unapproved workbasket,
if one does not exist it will create one. the changed objects within the workbasket
are shown with a "structure code" and "structure description" which are assumed to
be their identifying fields and descriptions (if existing).