TP-669: Remove manual WorkBasket assignment when editing.
Currently when editing a TrackedModel the user is asked which WorkBasket
to add the object to, and are offered to make a new WorkBasket if needed.
The intended usecase for now is to only allow a single WorkBasket at any
one time, in which case there is no reason to allow a choice or manual
creation.

This commit removes the interstitial page and defaults to using the current
active WorkBasket.