TP-709 Fix messages on wrapped business rule classes
It turns out that making a new subclass doesn't
cause most of the things we need to be inherited
by the new class, such as the Violation object and
the class' docstring. This results in cryptic
errors when the business rule fires. This new
method prevents that from happening by just
overriding the method.

As a bonus wrapped classes will now appear in
generated documentation.