Delete the Transition Period importers
The time has come. These served their purpose but
if we were to do this kind of exercise again, most
likely we'd do it in a very different way.

We've extracted most of the useful code from these
into other parts of the app, and we'll always have
these in git if we need to refer back to them.

The bits that are still sort of useful that we
should look to replace are:
- the `SheetImporter`, which would be useful to
  have back for adding things to prod that aren't
  in XML – but this isn't even the latest version
  as there is a better one on the fta branch.
- the `GoodsNomenclatureTreeCollector` which will
  come in handy for avoiding ME32 automatically,
  but again which will need some work.