More accurately describe XML types on parsers
This commit adds extra element types to XML parsers.
These more accurately describe the data types on XML
fields. This is currently only being used in the `clean`
methods to cast to an appropriate datatype but will
be used more extensively in future XML output work to
correctly format fields for output.