TP-647: Check users have permission to edit
This commit modifies our generic Create and Update class-based views to inherit from
PermissionRequiredMixin to ensure only users with permission to edit the tariff are
allowed to do so. This saves having to use the mixin with every component create and
update view class.

Also, a first pass at a generic BDD test for edit views is included.