Correctly order XML fields, including validity dates
Previously the XML parsers did not need to know
the order of the fields that they contained
because this ordering was not actually used. The
XML export code does use this though, so this
commit reorders the fields so that they are
listed in the same order in the class as they
appear in the XSD.

It's also necessary to include the validity fields
explicitly because they don't appear in
consistent places within the XML, so their
ordering needs to be recorded correctly.