TP-690: Model in use rules should respect deletes
Some business rules (particularly ones which prevent deleting a record) apply if a
record is currently "in use".  These rules did not correctly handle deleted records -
for example if a Footnote Type is associated with at least one Footnote record, then
that Footnote Type is considered "in use". But if all Footnotes that had been associated
with the type were then deleted, the Footnote Type would still consider itself "in use".
Similarly, if Footnote was subsequently associated with the type, but the transaction
containing that association was not yet approved, the Footnote Type would still consider
itself "in use".

This change corrects this class of business rules to correctly exclude deleted records
and those in "future" transactions from consideration of whether a record is "in use".