Prepend a sequential ID to generated messages.
The window function we use to get the message ID
generates one new ID per model. However, some
models (e.g. those with description fields) require
multiple message elements per model, which means
we would end up with duplicate message IDs.

The solution to this is to manually append a digit
to each message ID at SQL-generation time based on
the order in which the messages are output. This
ensures that the message IDs are unique.