TP-714: Endpoint that outputs our domain model schema.
Data Workspace are going to use our Activity
Stream endpoint to import data. But we need to
provide them with a list of the types we output
and the fields they should could/should store from
that data. This commit gets the system to output
this itself as an endpoint, so that it will stay
up to date as we evolve our schema.

DBML format is used, which is as good as anything,
and it should be possible for Data Workspace team
(as an offline step or otherwise) to transform
whatever this is into SQL CREATE TABLE statements
with the correct columns.