Add footnote applicability check to commodity codes
Only footnotes with a footnote type that have one
of three possible application codes are allowed
to be applied to commodity codes. This commit
implements the business rule NIG18 that ensures
only footnotes with the right type can be
associated.

It also DRYs up some similar logic for measures,
and makes what footnote codes should apply a
property of the models themselves.