Refactor importer tests to remove duplicated code
Our importer tests are quite verbose – in
particular the same tests are repeated once for
creation and again for delete/update. The two
tests also have different styles but are
functionally equivalent.

There is no real need for this, so this commit
removes this duplication. We now instead can
simply call `imported_fields_match` once and this
will test the importer for all variations of
update type.