XML tests should check that namespaces are correct
Our XML tests validate that the XML we output is
valid according to the TARIC3 spec and contain
the elements we expect. They do this by checking
for tags named correctly. They should also check
that the namespace is correct – ultimately, tags
that are named correctly but in the wrong
namespace are not what we want. This commit is a
find/replace to add the correct namespaces to all
of these checks.

We also don't need all of the fixture arguments
any more as these are handled automatically by
the `validate_taric_xml` decorator.