Remove validity setting on importer tests
The importer tests have previously had a parameter
that allowed setting validity dates on a test by
test basis. This is mostly unecessary because the
facotries will automatically set the validity
dates to appropriate values.

In most places this was used to swtich off
validity dates, which was unncessary since the
test logic did that anyway. In the remaining
places, it's not clear what this was really
achieveing – there isn't anywhere in the tariff
that should require an UPDATE or a DELETE to
change the validity dates because these two
mechanisms are independent. As evidecned, the
tests pass without this, and I cannot find any
reason for this feature to be preventing a
regression.

So this commit removes the test-by-test validity
date control. If we need to control this stuff,
it should really be done on the factories.