TP-788: Fix edits after publishing workbasket
After successfully editing and publishing a change, subsequent edits do not appear in
the list of tariff changes on the home page.

This is due to the workbasket not being updated in the user session. Subsequent edits
are appended to the already-published workbasket. The home page only lists the contents
of the current workbasket, which is a new and empty one after publishing the last one.

This change ensures that the session is updated on publishing the workbasket, and that
subsequent edits are correctly appended to the new unpublished workbasket.