Stop commodity codes being parents of themselves
There was a bug in the indents importer which
meant that a commodity code could end up as a
parent of itself if an indent is updated without
changing any of its fields. This was because the
parent search didn't correctly only allow the
item id to be equal in the case that the suffix
is also less.

This also fixes up some tests to use the import
infrastructure which makes it easy to swap out
our XML export mechanism later.