Switch to using Python-based querying and output
It turns out this is acceptably fast – about twice
as slow as having the CLI tools do all the work,
but given that this will run in an async task and
this method is far easier to test, this is
probably a good trade-off to make. It also makes
the code much simpler.