Fix reverse relations on all TrackedModels
When tracked models exist in a one-to-many
relationship (such as how one parent object can
have many descriptions) each description object
will hold a foreign key back to the parent object
that specifically links to the version it was
created against.

    Parent X (primary key = 1)
    Description X (primary key = 2, parent = 1)

Calling the reverse relation on the parent will
correctly discover the descriptions:

    parent.descriptions
    # => [<Description: pk=2>]

If that parent object is updated, it receives a
new row in the database, but the descriptions are
not updated. This means that the reverse
relationship now no longer returns any objects,
because the reverse relationship by default tries
to find descriptions that have the primary key of
the *new* parent object:

    parent_v2 = parent.new_version(...)
    parent_v2.descriptions # => []

What needs to be done instead is not use the
database primary key of the parent object but
instead refer to that object's version group,
which will stay constant between versions.

    Parent.objects = TrackedModelManager.from_queryset(...)
    parent_v2.descriptions # => [<Description: pk=2>]

This commit therefore creates a new Manager class
(that will be inherited by RelatedManager) and
that replaces PKs with the version group of the
instance. For more information on why this works
see django/db/models/fields/related_descriptors.py.